Tree-drawing improvements
I put a class wrapper around the tree-drawing logic to unify parameters
... but it turned out that it didn't matter, because the problem was a
data import problem. Maybe the work will be useful in the future.

At the same time, I discovered the 'group' attribute for nodes, and
especially since non-cluster subgraphs seem to be a little broken,
this has finally allowed me to clean up some of the more spaghetti bits
that tended to happen from time to time.