Push image before end of the cloud build trigger.
This is an absolutely fascinating set of weirdnesses.
First, apparently the Google docker image doesn't support build --push.
Second, build rules will push any tags in the "images" section, but only
when the entire build is complete. So you actually need to do a manual
push to both build an image and deploy it, because you're not allowed
to do it automatically except in cases where it happens after you're done.