Small edits to make build verification pass
Looking back at this, I noticed that while all steps succeeded,
Cloud Build was reporting that the overall build was failing.
I think that's because I was relying on an attribute which was for
verification to instead use to push a tag, because that had worked once
before when I was only building and not deploying in the build trigger.
So, now that I understand Docker a bit better, I'm tagging it properly.

Also, while I'm here, I'm cleaning up some of the weirder YAML syntax
that I copied from examples I saw along the way.